+++ components/MarkdownEditor.tsx
@@ -1,7 +1,13 @@
 import "@uiw/react-md-editor/markdown-editor.css";
 import "@uiw/react-markdown-preview/markdown.css";
 
-import { Button, Icon, Label, Tooltip } from "@equinor/eds-core-react";
+import {
+  Button,
+  Icon,
+  Label,
+  Tooltip,
+  Typography,
+} from "@equinor/eds-core-react";
 import MDEditor, { ICommand, TextState } from "@uiw/react-md-editor";
 import { useEffect, useState } from "react";
 
@@ -19,8 +25,11 @@ export default function MarkdownEditor(props: {
   defaultText: string;
   label: string;
   onChange?: (value?: string) => void;
+  helperText?: string;
+  requireText?: boolean;
 }) {
-  const { canEdit, defaultText, label, onChange } = props;
+  const { canEdit, defaultText, label, onChange, helperText, requireText } =
+    props;
   const [editMode, setEditMode] = useState(false);
   const [isOpenUrlPrompt, setIsOpenUrlPrompt] = useState(false);
   const [selectionInfo, setSelectionInfo] = useState<SelectionInfo>({
@@ -30,6 +39,7 @@ export default function MarkdownEditor(props: {
   });
   // The "text"-state is used for displaying the changing text instantly
   const [text, setText] = useState<string | undefined>(defaultText);
+  const missingText = requireText && text?.length === 0;
 
   useEffect(() => {
     setText(defaultText);
@@ -69,9 +79,7 @@ export default function MarkdownEditor(props: {
   // Sets local state to display text instantly and updates the API
   const setAndPatchText = (text?: string) => {
     // Patching the text via the API is debounced in the onChange
-    if (onChange) {
-      onChange(text);
-    }
+    onChange && onChange(text);
     setText(text);
   };
 
@@ -113,6 +121,7 @@ export default function MarkdownEditor(props: {
       <div style={{ display: "flex", gap: 12 }}>
         <div onClick={() => canEdit && setEditMode(true)} style={{ flex: 1 }}>
           {/* Override MDEditor default styles */}
+          {/* eslint-disable-next-line react/no-unknown-property */}
           <style global jsx>{`
             #mdEditor > div.w-md-editor-content > div > div {
               font-size: 1rem;
@@ -155,13 +164,20 @@ export default function MarkdownEditor(props: {
             }}
             style={{
               border: editMode
-                ? `2px solid ${colors.EQUINOR_PROMINENT}`
+                ? `2px solid ${
+                    missingText ? colors.DANGER : colors.EQUINOR_PROMINENT
+                  }`
                 : `1px solid ${colors.EQUINOR_DISABLED}`,
               borderRadius: 0,
               boxShadow: canEdit && !editMode ? "0 1px 0 0 gray" : "none",
             }}
             autoFocus
           />
+          {missingText && (
+            <Typography color="danger" variant="caption">
+              Description is required
+            </Typography>
+          )}
         </div>
         {editMode && (
           <Button
@@ -170,11 +186,22 @@ export default function MarkdownEditor(props: {
               minWidth: 48,
             }}
             variant="ghost_icon"
+            disabled={missingText}
           >
-            <Icon data={check} color={colors.EQUINOR_PROMINENT} />
+            <Icon
+              data={check}
+              color={
+                missingText ? colors.EQUINOR_DISABLED : colors.EQUINOR_PROMINENT
+              }
+            />
           </Button>
         )}
       </div>
+      {helperText && (
+        <Typography style={{ marginTop: 12 }} variant="caption">
+          {helperText}
+        </Typography>
+      )}
     </div>
   );
 }