+++ components/canvas/Canvas.tsx
@@ -22,18 +22,15 @@ import { LiveIndicator } from "../LiveIndicator";
 import { SideBar } from "../SideBar";
 import styles from "./Canvas.module.scss";
 import { nodeElementTypes } from "./NodeElementTypes";
-import { OldFlytButton } from "./OldFlytButton";
 import { ToBeToggle } from "./ToBeToggle";
 import { useAccess } from "./hooks/useAccess";
 import { useCenterCanvas } from "./hooks/useCenterCanvas";
 import { setLayout } from "./hooks/useLayout";
-import { useNodeAdd } from "./hooks/useNodeAdd";
 import { useNodeDrag } from "./hooks/useNodeDrag";
 import { useNodeMerge } from "./hooks/useNodeMerge";
 import { useWebSocket } from "./hooks/useWebSocket";
 import { getQIPRContainerWidth } from "./utils/getQIPRContainerWidth";
 import { setMainActivitiesDurationSum } from "./utils/setMainActivitiesDurationSum";
-import { useProjectId } from "@/hooks/useProjectId";
 import { useEdgeDelete } from "./hooks/useEdgeDelete";
 import { ScrimDelete } from "../ScrimDelete";
 import { MiniMapCustom } from "@/components/canvas/MiniMapCustom";
@@ -54,13 +51,9 @@ const Canvas = ({
   const [selectedNode, setSelectedNode] = useState<Node<NodeData> | undefined>(
     undefined
   );
-  const { projectId } = useProjectId();
   const { userCanEdit } = useAccess(project);
 
   const shapeSize = { height: 140, width: 140 };
-  const createdBeforeSecondMajorRelease =
-    new Date(project.created).getTime() <
-    new Date("2024-04-24T00:08:00.000000Z").getTime();
 
   let tempNodes: Node<NodeData>[] = [];
   const tempEdges: Edge[] = apiEdges.map((e) => ({ ...e, label: e.edgeValue }));
@@ -78,7 +71,6 @@ const Canvas = ({
 
   const { onNodeDragStart, onNodeDrag, onNodeDragStop } = useNodeDrag();
   const { mutate: mergeNode, merging } = useNodeMerge();
-  const { mutate: addNode } = useNodeAdd();
   const { deleteEdgeMutation } = useEdgeDelete();
 
   const { socketConnected, socketReason } = useWebSocket();
@@ -129,8 +121,6 @@ const Canvas = ({
         data: {
           ...node,
           handleClickNode: () => handleSetSelectedNode(node.id),
-          handleClickAddNode: (id, type, position) =>
-            addNode({ parentId: id, type, position }),
           handleMerge: (sourceId, targetId) =>
             sourceId && targetId && mergeNode.mutate({ sourceId, targetId }),
           mergeable:
@@ -353,6 +343,7 @@ const Canvas = ({
         onEdgeMouseLeave={() => handleSetSelectedEdge(undefined)}
         attributionPosition="bottom-right"
         connectionRadius={100}
+        nodeDragThreshold={15}
       >
         <MiniMapCustom />
         <Controls className={styles.controls} showInteractive={false}>
@@ -361,9 +352,6 @@ const Canvas = ({
           </ControlButton>
         </Controls>
       </ReactFlow>
-      {createdBeforeSecondMajorRelease && (
-        <OldFlytButton projectId={projectId} />
-      )}
     </>
   );
 };