+++ types/NodeData.ts
@@ -1,6 +1,4 @@
-import { Position } from "reactflow";
 import { NodeDataApi } from "./NodeDataApi";
-import { NodeTypes } from "./NodeTypes";
 import { timeDefinitions } from "../utils/unitDefinitions";
 
 export type NodeData = {
@@ -11,11 +9,6 @@ export type NodeData = {
   mergeOption?: boolean;
   handleMerge?: (sourceId: string | null, targetId: string | null) => void;
   handleClickNode?: () => void;
-  handleClickAddNode?: (
-    nodeId: string,
-    type: NodeTypes,
-    position: Position
-  ) => void;
   mergeable?: boolean;
   merging?: boolean;
   userCanEdit?: boolean;